+++ connectionpool.go
@@ -187,8 +187,9 @@ func connConfig(cfg *ClusterConfig) (*ConnConfig, error) {
 		}
 
 		hostDialer = &defaultHostDialer{
-			dialer:    dialer,
-			tlsConfig: tlsConfig,
+			dialer:         dialer,
+			tlsConfig:      tlsConfig,
+			connectTimeout: cfg.ConnectTimeout,
 		}
 	}